Refactored input validators into `tui\utils\tui_validators.py`
